refactor: split grid column resizing hook
