+++ src/components/grid-layout/hooks/constants.ts
@@ -1,0 +1,2 @@
+export const MIN_COLUMN_WIDTH = 200;
+export const COLLAPSED_WIDTH = 44;
